others/requirements.txt
torchtext==0.6.0
pandas==1.0.0
matplotlib==3.1.2
seaborn==0.10.0
plotly==4.5.2
plotly-express==0.4.1
dash==1.7.0
jupyter==1.0.0
jupyterlab==1.2.3
ipywidgets>=7.5
scikit-learn==0.22
beautifultable==0.7.0
dash-core-components==1.6.0
torchvision
pyyaml
git+https://github.com/ildoonet/pytorch-randaugment